[back] refactor: split serializers into several files
