--- backend/tournesol/query_serializers.py
@@ -1,14 +1,0 @@
-from rest_framework import serializers
-
-
-class RecommendationsFilterSerializer(serializers.Serializer):
-    date_lte = serializers.DateTimeField(default=None)
-    date_gte = serializers.DateTimeField(default=None)
-    search = serializers.CharField(
-        default=None, help_text="A search query to filter entities"
-    )
-    unsafe = serializers.BooleanField(
-        default=False,
-        help_text="If true, entities considered as unsafe recommendations because of a"
-        " low score or due to too few contributions will be included.",
-    )
