--- backend/tournesol/serializers.py
@@ -1,377 +1,0 @@
-"""
-Serializer used by Tournesol's API
-"""
-
-from typing import Optional
-
-from django.db import IntegrityError, transaction
-from django.db.models import ObjectDoesNotExist, Q
-from drf_spectacular.types import OpenApiTypes
-from drf_spectacular.utils import extend_schema_field
-from rest_framework import serializers
-from rest_framework.exceptions import NotFound, ValidationError
-from rest_framework.fields import BooleanField, IntegerField, RegexField, SerializerMethodField
-from rest_framework.serializers import ModelSerializer, Serializer
-
-from core.utils.constants import YOUTUBE_VIDEO_ID_REGEX
-from tournesol.errors import ConflictError
-from tournesol.utils.api_youtube import VideoNotFound
-
-from .models import (
-    Comparison,
-    ComparisonCriteriaScore,
-    ContributorRating,
-    ContributorRatingCriteriaScore,
-    CriteriaRank,
-    Entity,
-    EntityCriteriaScore,
-    Poll,
-    VideoRateLater,
-)
-
-
-class VideoSerializer(ModelSerializer):
-    duration = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Entity
-        fields = [
-            "video_id",
-            "name",
-            "description",
-            "publication_date",
-            "views",
-            "uploader",
-            "language",
-            "rating_n_ratings",
-            "rating_n_contributors",
-            "duration",
-        ]
-        read_only_fields = [
-            "name",
-            "description",
-            "publication_date",
-            "views",
-            "uploader",
-            "language",
-            "rating_n_ratings",
-            "rating_n_contributors",
-            "duration",
-        ]
-
-    def create(self, validated_data):
-        try:
-            return Entity.create_from_video_id(validated_data["video_id"])
-        except VideoNotFound:
-            raise NotFound("The video has not been found. `video_id` may be incorrect.")
-
-    # Convert duration to seconds to facilitate use of Humanize package
-    def get_duration(self, obj) -> Optional[int]:
-        if obj.duration:
-            return int(obj.duration.total_seconds())
-        return None
-
-
-class RelatedVideoSerializer(VideoSerializer):
-    """
-    A video serializer that will create the Entity object on validation
-    if it does not exist in the database yet.
-
-    Used by ModelSerializer(s) having one or more nested relations with Entity,
-    and having the constraint to ensure that video instances exist before
-    they can be saved properly.
-    """
-
-    video_id = RegexField(YOUTUBE_VIDEO_ID_REGEX)
-
-    def validate_video_id(self, value):
-        try:
-            Entity.objects.get(video_id=value)
-        except ObjectDoesNotExist:
-            try:
-                Entity.create_from_video_id(value)
-            except VideoNotFound:
-                raise ValidationError("The video has not been found. `video_id` may be incorrect.")
-        return value
-
-
-class EntityCriteriaScoreSerializer(ModelSerializer):
-    class Meta:
-        model = EntityCriteriaScore
-        fields = ["criteria", "score"]
-
-
-class VideoSerializerWithCriteria(VideoSerializer):
-    criteria_scores = EntityCriteriaScoreSerializer(many=True)
-
-    class Meta(VideoSerializer.Meta):
-        fields = VideoSerializer.Meta.fields + ['criteria_scores']
-        # Overriding 'read_only_fields' should not be necessary here. However, due to a
-        # limitation in DRF, some extra attributes from model fields (nullable, etc.) are
-        # not present in the schema generated by drf-spectacular for read_only fields.
-        # See https://github.com/tfranzel/drf-spectacular/issues/383 for more details.
-        # This serializer is always used as read-only, so the 'read_only_fields' definition
-        # can be discarded safely to generate a correct OpenAPI schema.
-        read_only_fields = []
-
-
-class VideoRateLaterSerializer(ModelSerializer):
-    video = RelatedVideoSerializer()
-    user = serializers.HiddenField(default=serializers.CurrentUserDefault())
-
-    class Meta:
-        model = VideoRateLater
-        fields = ["user", "video"]
-
-    def create(self, validated_data):
-        video_id = validated_data.pop("video")["video_id"]
-        video = Entity.objects.get(video_id=video_id)
-        try:
-            return super().create({"video": video, **validated_data})
-        except IntegrityError:
-            raise ConflictError
-
-
-class ComparisonCriteriaScoreSerializer(ModelSerializer):
-    class Meta:
-        model = ComparisonCriteriaScore
-        fields = ["criteria", "score", "weight"]
-
-    def validate_criteria(self, value):
-        current_poll = self.context["poll"]
-        if value not in current_poll.criterias_list:
-            raise ValidationError(
-                f"'{value}' is not a valid criteria for poll '{current_poll.name}'"
-            )
-        return value
-
-
-class ComparisonSerializerMixin:
-    def reverse_criteria_scores(self, criteria_scores):
-        opposite_scores = criteria_scores.copy()
-        for index, score in enumerate(criteria_scores):
-            opposite_scores[index]["score"] = score["score"] * -1
-
-        return opposite_scores
-
-
-class ComparisonSerializer(ComparisonSerializerMixin, ModelSerializer):
-    """
-    A comparison serializer used for the operations create, retrieve and list.
-
-    Given the context determined by the view, it will represent the comparison
-    in the reverse order.
-
-    Use `ComparisonUpdateSerializer` for the update operation.
-    """
-
-    entity_a = RelatedVideoSerializer(source="entity_1")
-    entity_b = RelatedVideoSerializer(source="entity_2")
-    criteria_scores = ComparisonCriteriaScoreSerializer(many=True)
-    user = serializers.HiddenField(default=serializers.CurrentUserDefault())
-
-    class Meta:
-        model = Comparison
-        fields = ["user", "entity_a", "entity_b", "criteria_scores", "duration_ms"]
-
-    def to_representation(self, instance):
-        """
-        Display the opposite of each criteria scores if the comparison is
-        requested in the reverse order.
-        """
-        ret = super(ComparisonSerializer, self).to_representation(instance)
-
-        if self.context.get("reverse", False):
-            ret["entity_a"], ret["entity_b"] = ret["entity_b"], ret["entity_a"]
-            ret["criteria_scores"] = self.reverse_criteria_scores(
-                ret["criteria_scores"]
-            )
-
-        return ret
-
-    @transaction.atomic
-    def create(self, validated_data):
-        video_id_1 = validated_data.pop("entity_1").get("video_id")
-        video_id_2 = validated_data.pop("entity_2").get("video_id")
-        # the validation performed by the RelatedVideoSerializer guarantees
-        # that the videos submitted exist in the database
-        video_1 = Entity.objects.get(video_id=video_id_1)
-        video_2 = Entity.objects.get(video_id=video_id_2)
-        criteria_scores = validated_data.pop("criteria_scores")
-
-        comparison = Comparison.objects.create(
-            poll=self.context.get('poll'),
-            entity_1=video_1,
-            entity_2=video_2,
-            **validated_data,
-        )
-
-        for criteria_score in criteria_scores:
-            ComparisonCriteriaScore.objects.create(
-                comparison=comparison, **criteria_score
-            )
-
-        return comparison
-
-
-class ComparisonUpdateSerializer(ComparisonSerializerMixin, ModelSerializer):
-    """
-    A comparison serializer used only for updates.
-
-    Given the context determined by the view, it will represent or save the
-    comparison in the reverse order.
-
-    Use `ComparisonSerializer` for all other operations.
-    """
-
-    criteria_scores = ComparisonCriteriaScoreSerializer(many=True)
-    entity_a = VideoSerializer(source="entity_1", read_only=True)
-    entity_b = VideoSerializer(source="entity_2", read_only=True)
-
-    class Meta:
-        model = Comparison
-        fields = ["criteria_scores", "duration_ms", "entity_a", "entity_b"]
-
-    def to_representation(self, instance):
-        """
-        Display the opposite of each criteria scores if the comparison is
-        requested in the reverse order.
-
-        Also add `entity_a` and `entity_b` fields to make the representation
-        consistent across all comparison serializers.
-        """
-        ret = super(ComparisonUpdateSerializer, self).to_representation(instance)
-
-        if self.context.get("reverse", False):
-            ret["entity_a"], ret["entity_b"] = ret["entity_b"], ret["entity_a"]
-            ret["criteria_scores"] = self.reverse_criteria_scores(
-                ret["criteria_scores"]
-            )
-
-        ret.move_to_end("entity_b", last=False)
-        ret.move_to_end("entity_a", last=False)
-        return ret
-
-    def to_internal_value(self, data):
-        """
-        Save the comparison in the order expected by the model, even if the
-        comparison is provided reversed.
-        """
-        ret = super(ComparisonUpdateSerializer, self).to_internal_value(data)
-
-        if self.context.get("reverse", False):
-            ret["criteria_scores"] = self.reverse_criteria_scores(
-                ret["criteria_scores"]
-            )
-        return ret
-
-    @transaction.atomic
-    def update(self, instance, validated_data):
-        if validated_data.get("duration_ms"):
-            instance.duration_ms = validated_data.get("duration_ms")
-
-        instance.save()
-        instance.criteria_scores.all().delete()
-
-        for criteria_score in validated_data.pop("criteria_scores"):
-            instance.criteria_scores.create(**criteria_score)
-
-        return instance
-
-
-class ContributorCriteriaScore(ModelSerializer):
-    class Meta:
-        model = ContributorRatingCriteriaScore
-        fields = ["criteria", "score", "uncertainty"]
-
-
-class ContributorRatingSerializer(ModelSerializer):
-    video = VideoSerializer(source="entity", read_only=True)
-    criteria_scores = ContributorCriteriaScore(many=True, read_only=True)
-    n_comparisons = SerializerMethodField(
-        help_text="Number of comparisons submitted by the current user about the current video",
-    )
-
-    class Meta:
-        model = ContributorRating
-        fields = ["video", "is_public", "criteria_scores", "n_comparisons"]
-
-    @extend_schema_field(OpenApiTypes.INT)
-    def get_n_comparisons(self, obj):
-        if hasattr(obj, "n_comparisons"):
-            # Use annotated field if it has been defined by the queryset
-            return obj.n_comparisons
-        return obj.user.comparisons.filter(
-            Q(entity_1=obj.entity) | Q(entity_2=obj.entity)
-        ).count()
-
-
-class ContributorRatingCreateSerializer(ContributorRatingSerializer):
-    video_id = RegexField(YOUTUBE_VIDEO_ID_REGEX, write_only=True)
-
-    class Meta:
-        model = ContributorRating
-        fields = ["video_id", "is_public", "video", "criteria_scores", "n_comparisons"]
-
-    def validate(self, attrs):
-        video_id = attrs.pop("video_id")
-        video_serializer = RelatedVideoSerializer(data={"video_id": video_id})
-        video_serializer.is_valid(raise_exception=True)
-        video = Entity.objects.get(video_id=video_id)
-        user = self.context["request"].user
-        if user.contributorvideoratings.filter(entity=video).exists():
-            raise ValidationError(
-                "A ContributorRating already exists for this (user, video)",
-                code='unique',
-            )
-        attrs["entity"] = video
-        attrs["user"] = user
-        return attrs
-
-
-class ContributorRatingUpdateAllSerializer(Serializer):
-    is_public = BooleanField()
-
-
-class StatisticsSerializer(Serializer):
-    user_count = IntegerField()
-    last_month_user_count = IntegerField()
-    video_count = IntegerField()
-    last_month_video_count = IntegerField()
-    comparison_count = IntegerField()
-    last_month_comparison_count = IntegerField()
-
-
-class PollCriteriaSerializer(ModelSerializer):
-    name = serializers.CharField(source="criteria.name")
-    label = serializers.CharField(source="criteria.get_label")
-
-    class Meta:
-        model = CriteriaRank
-        fields = ["name", "label", "optional"]
-
-
-class PollSerializer(ModelSerializer):
-    criterias = PollCriteriaSerializer(source="criteriarank_set", many=True)
-
-    class Meta:
-        model = Poll
-        fields = ["name", "criterias"]
-
-
-class RecommendationSerializer(ModelSerializer):
-    n_comparisons = serializers.IntegerField(source="rating_n_ratings")
-    n_contributors = serializers.IntegerField(source="rating_n_contributors")
-    criteria_scores = EntityCriteriaScoreSerializer(many=True)
-    total_score = serializers.FloatField()
-
-    class Meta:
-        model = Entity
-        fields = [
-            "uid",
-            "n_comparisons",
-            "n_contributors",
-            "metadata",
-            "total_score",
-            "criteria_scores",
-        ]
