+++ backend/tournesol/views/exports.py
@@ -9,7 +9,7 @@ from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.views import APIView
 
 from tournesol.models import Comparison, ContributorRating
-from tournesol.serializers import ComparisonSerializer
+from tournesol.serializers.comparison import ComparisonSerializer
 from tournesol.utils.cache import cache_page_no_i18n
 
 
@@ -18,22 +18,24 @@ def write_comparisons_file(request, write_target):
     Writes a user's comparisons as a CSV file to write_target which can be
     among other options an HttpResponse or a StringIO
     """
-    fieldnames = ['video_a', 'video_b', 'criteria', 'weight', 'score']
+    fieldnames = ["video_a", "video_b", "criteria", "weight", "score"]
     writer = csv.DictWriter(write_target, fieldnames=fieldnames)
     writer.writeheader()
-    comparisons = Comparison.objects.filter(user=request.user)\
-        .select_related("entity_1", "entity_2")\
+    comparisons = (
+        Comparison.objects.filter(user=request.user)
+        .select_related("entity_1", "entity_2")
         .prefetch_related("criteria_scores")
+    )
     serialized_comparisons = ComparisonSerializer(comparisons, many=True).data
 
     writer.writerows(
         {
             "video_a": comparison["entity_a"]["video_id"],
             "video_b": comparison["entity_b"]["video_id"],
-            **criteria_score
+            **criteria_score,
         }
         for comparison in serialized_comparisons
-        for criteria_score in comparison['criteria_scores']
+        for criteria_score in comparison["criteria_scores"]
     )
 
 
@@ -42,19 +44,34 @@ def write_public_comparisons_file(request, write_target):
     Writes all public comparisons data as a CSV file to write_target which can be
     among other options an HttpResponse or a StringIO
     """
-    fieldnames = ['public_username', 'video_a', 'video_b', 'criteria', 'weight', 'score']
+    fieldnames = [
+        "public_username",
+        "video_a",
+        "video_b",
+        "criteria",
+        "weight",
+        "score",
+    ]
     writer = csv.DictWriter(write_target, fieldnames=fieldnames)
     writer.writeheader()
-    public_data = ContributorRating.objects.filter(is_public=True).select_related("user", "entity")
+    public_data = ContributorRating.objects.filter(is_public=True).select_related(
+        "user", "entity"
+    )
     serialized_comparisons = []
     public_videos = set((rating.user, rating.entity) for rating in public_data)
-    comparisons = Comparison.objects.all()\
-        .select_related("entity_1", "entity_2", "user")\
+    comparisons = (
+        Comparison.objects.all()
+        .select_related("entity_1", "entity_2", "user")
         .prefetch_related("criteria_scores")
+    )
     public_comparisons = [
-        comparison for comparison in comparisons
-        if ((comparison.user, comparison.entity_1) in public_videos
-            and (comparison.user, comparison.entity_2) in public_videos)]
+        comparison
+        for comparison in comparisons
+        if (
+            (comparison.user, comparison.entity_1) in public_videos
+            and (comparison.user, comparison.entity_2) in public_videos
+        )
+    ]
     public_usernames = [comparison.user.username for comparison in public_comparisons]
     serialized_comparisons = ComparisonSerializer(public_comparisons, many=True).data
     writer.writerows(
@@ -62,10 +79,12 @@ def write_public_comparisons_file(request, write_target):
             "public_username": public_username,
             "video_a": comparison["entity_a"]["video_id"],
             "video_b": comparison["entity_b"]["video_id"],
-            **criteria_score
+            **criteria_score,
         }
-        for (public_username, comparison) in zip(public_usernames, serialized_comparisons)
-        for criteria_score in comparison['criteria_scores']
+        for (public_username, comparison) in zip(
+            public_usernames, serialized_comparisons
+        )
+        for criteria_score in comparison["criteria_scores"]
     )
 
 
@@ -75,11 +94,11 @@ class ExportComparisonsView(APIView):
 
     @extend_schema(
         description="Download current user data in .zip file",
-        responses={200: OpenApiTypes.BINARY}
+        responses={200: OpenApiTypes.BINARY},
     )
     def get(self, request):
-        response = HttpResponse(content_type='text/csv')
-        response['Content-Disposition'] = 'attachment; filename="export.csv"'
+        response = HttpResponse(content_type="text/csv")
+        response["Content-Disposition"] = 'attachment; filename="export.csv"'
         write_comparisons_file(request, response)
         return response
 
@@ -88,14 +107,16 @@ class ExportPublicComparisonsView(APIView):
     permission_classes = [AllowAny]
     throttle_scope = "api_export_comparisons"
 
-    @method_decorator(cache_page_no_i18n(60*10))  # 10 minutes cache
+    @method_decorator(cache_page_no_i18n(60 * 10))  # 10 minutes cache
     @extend_schema(
         description="Download public data in .csv file",
-        responses={200: OpenApiTypes.BINARY}
+        responses={200: OpenApiTypes.BINARY},
     )
     def get(self, request):
-        response = HttpResponse(content_type='text/csv')
-        response['Content-Disposition'] = 'attachment; filename="tournesol_public_export.csv"'
+        response = HttpResponse(content_type="text/csv")
+        response[
+            "Content-Disposition"
+        ] = 'attachment; filename="tournesol_public_export.csv"'
         write_public_comparisons_file(request, response)
         return response
 
@@ -105,14 +126,14 @@ class ExportAllView(APIView):
 
     @extend_schema(
         description="Download current user data in .zip file",
-        responses={200: OpenApiTypes.BINARY}
+        responses={200: OpenApiTypes.BINARY},
     )
     def get(self, request):
         # Folder name in ZIP archive which contains the above files
         zip_root = f"export_{request.user.username}"
 
         response = HttpResponse(content_type="application/x-zip-compressed")
-        response['Content-Disposition'] = f'attachment; filename={zip_root}.zip'
+        response["Content-Disposition"] = f"attachment; filename={zip_root}.zip"
 
         zf = zipfile.ZipFile(response, "w", compression=zipfile.ZIP_DEFLATED)
 
@@ -120,7 +141,7 @@ class ExportAllView(APIView):
         # content of the export in the future
         with StringIO() as comparisons_file_like:
             write_comparisons_file(request, comparisons_file_like)
-            zf.writestr(f'{zip_root}/comparisons.csv', comparisons_file_like.getvalue())
+            zf.writestr(f"{zip_root}/comparisons.csv", comparisons_file_like.getvalue())
 
         # Close zip for all contents to be written
         zf.close()
