+++ backend/tournesol/views/video.py
@@ -14,6 +14,9 @@ from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.permissions import IsAuthenticatedOrReadOnly
 from rest_framework.viewsets import GenericViewSet
 
+from tournesol.entities import VideoEntity
+from tournesol.models import Entity
+from tournesol.serializers.entity import VideoSerializer, VideoSerializerWithCriteria
 from tournesol.throttling import (
     BurstAnonRateThrottle,
     BurstUserRateThrottle,
@@ -22,61 +25,59 @@ from tournesol.throttling import (
     SustainedUserRateThrottle,
 )
 
-from ..entities import VideoEntity
-from ..models import Entity
-from ..serializers import VideoSerializer, VideoSerializerWithCriteria
-
 
 @extend_schema_view(
     create=extend_schema(
         description="Add a video to the db if it does not already exist."
     ),
-    retrieve=extend_schema(
-        description="Retrieve details about a single video."
-    ),
+    retrieve=extend_schema(description="Retrieve details about a single video."),
     list=extend_schema(
         description="Retrieve a list of recommended videos, sorted by decreasing total score.",
         parameters=[
             OpenApiParameter("search"),
-            OpenApiParameter("language",
-                             description="Accepted languages separated by commas "
-                                         "(e.g. 'en,fr,de'). If empty, accept all languages."),
+            OpenApiParameter(
+                "language",
+                description="Accepted languages separated by commas "
+                "(e.g. 'en,fr,de'). If empty, accept all languages.",
+            ),
             OpenApiParameter("uploader"),
             OpenApiParameter(
                 "date_lte",
                 OpenApiTypes.DATETIME,
                 description="Return videos published **before** this date.  \n"
                 "Accepted formats: ISO 8601 datetime (e.g `2021-12-01T12:45:00`) "
-                "or legacy: `dd-mm-yy-hh-mm-ss`."
+                "or legacy: `dd-mm-yy-hh-mm-ss`.",
             ),
             OpenApiParameter(
                 "date_gte",
                 OpenApiTypes.DATETIME,
                 description="Return videos published **after** this date.  \n"
                 "Accepted formats: ISO 8601 datetime (e.g `2021-12-01T12:45:00`) "
-                "or legacy: `dd-mm-yy-hh-mm-ss`."
+                "or legacy: `dd-mm-yy-hh-mm-ss`.",
             ),
             OpenApiParameter(
                 "unsafe",
                 OpenApiTypes.BOOL,
                 description="If true, videos considered as unsafe recommendations because of a "
-                "low score or due to too few contributions will be included."
+                "low score or due to too few contributions will be included.",
             ),
             *[
                 OpenApiParameter(
                     crit,
                     OpenApiTypes.INT,
-                    description=f"Weight for criteria '{crit}', between 0 and 100"
+                    description=f"Weight for criteria '{crit}', between 0 and 100",
                 )
                 for crit in settings.LEGACY_CRITERIAS
             ],
         ],
-    )
+    ),
 )
-class VideoViewSet(mixins.CreateModelMixin,
-                   mixins.RetrieveModelMixin,
-                   mixins.ListModelMixin,
-                   GenericViewSet):
+class VideoViewSet(
+    mixins.CreateModelMixin,
+    mixins.RetrieveModelMixin,
+    mixins.ListModelMixin,
+    GenericViewSet,
+):
     queryset = Entity.objects.all()
     pagination_class = LimitOffsetPagination
     permission_classes = [IsAuthenticatedOrReadOnly]
@@ -86,7 +87,7 @@ class VideoViewSet(mixins.CreateModelMixin,
         BurstAnonRateThrottle,
         BurstUserRateThrottle,
         SustainedAnonRateThrottle,
-        SustainedUserRateThrottle
+        SustainedUserRateThrottle,
     ]
     throttle_scope = "api_video_post"
 
@@ -97,8 +98,8 @@ class VideoViewSet(mixins.CreateModelMixin,
         Parse ISO datetime from query string.
         Also accepts legacy format 'DD-MM-YY-HH-MM-SS'
         """
-        if re.match(r'^\d{2}-\d{2}-\d{2}-\d{2}-\d{2}-\d{2}$', value):
-            return timezone.datetime.strptime(value, '%d-%m-%y-%H-%M-%S')
+        if re.match(r"^\d{2}-\d{2}-\d{2}-\d{2}-\d{2}-\d{2}$", value):
+            return timezone.datetime.strptime(value, "%d-%m-%y-%H-%M-%S")
         parsed = dateparse.parse_datetime(value)
         if parsed is None:
             raise ValueError(f'Failed to parse "{value}" as datetime')
@@ -111,22 +112,22 @@ class VideoViewSet(mixins.CreateModelMixin,
         request = self.request
         queryset = self.queryset
 
-        uploader = request.query_params.get('uploader')
+        uploader = request.query_params.get("uploader")
         if uploader:
             queryset = queryset.filter(uploader=uploader)
 
-        search = request.query_params.get('search')
+        search = request.query_params.get("search")
         if search:
             queryset = VideoEntity.filter_search(queryset, search)
 
-        date_lte = request.query_params.get('date_lte') or ""
+        date_lte = request.query_params.get("date_lte") or ""
         if date_lte:
             try:
                 date_lte = self.parse_datetime(date_lte)
                 queryset = VideoEntity.filter_date_lte(queryset, date_lte)
             except ValueError:
                 raise ValidationError('"date_lte" is an invalid datetime.')
-        date_gte = request.query_params.get('date_gte') or ""
+        date_gte = request.query_params.get("date_gte") or ""
         if date_gte:
             try:
                 date_gte = self.parse_datetime(date_gte)
@@ -156,25 +157,16 @@ class VideoViewSet(mixins.CreateModelMixin,
             total_score=Sum(F("criteria_scores__score") * criteria_weight)
         )
 
-        show_unsafe = request.query_params.get('unsafe') == 'true'
+        show_unsafe = request.query_params.get("unsafe") == "true"
 
         if show_unsafe is True:
-            queryset = (
-                queryset
-                .filter(total_score__isnull=False)
-            )
+            queryset = queryset.filter(total_score__isnull=False)
         else:
-            queryset = (
-                queryset
-                .filter(
-                    rating_n_contributors__gte=settings.RECOMMENDATIONS_MIN_CONTRIBUTORS
-                )
-                .filter(total_score__gt=0)
-            )
-        return (
-            queryset
-            .prefetch_related("criteria_scores")
-            .order_by('-total_score', '-publication_date')
+            queryset = queryset.filter(
+                rating_n_contributors__gte=settings.RECOMMENDATIONS_MIN_CONTRIBUTORS
+            ).filter(total_score__gt=0)
+        return queryset.prefetch_related("criteria_scores").order_by(
+            "-total_score", "-publication_date"
         )
 
     def get_serializer_class(self):
